requirements.txt
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.14
langchain-openai==0.0.7
chromadb==0.4.24